requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.5.7
chainer==7.8.1
chardet==5.2.0
charset-normalizer==3.1.0
cmake==3.26.4
contourpy==1.0.7
cycler==0.11.0
Deprecated==1.2.14
filelock==3.12.2
fonttools==4.39.4
frozenlist==1.4.0
future==0.18.3
Hexy==1.4.4
htm.core==2.1.16
idna==3.4
iniconfig==2.0.0
joblib==1.3.2
jsonpickle==3.0.2
kiwisolver==1.4.4
line-bot-sdk==3.5.0
matplotlib==3.7.1
mock==5.0.2
more-itertools==10.1.0
multidict==6.0.4
music21==9.1.0
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
pluggy==1.1.0
ply==3.11
prettytable==3.8.0
protobuf==4.23.3
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.0.9
PySimpleGUI==4.60.5
pytest==7.3.2
python-dateutil==2.8.2
requests==2.31.0
scipy==1.10.1
sip==6.7.11
six==1.16.0
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
wcwidth==0.2.6
webcolors==1.13
wrapt==1.15.0
yarl==1.9.2
